requirements.txt
django-mssql-backend
azure-identity
azure-keyvault-secrets
pyodbc
cryptography
optimus_ids 